@@ -1,58 +1,31 @@
-[tool.poetry]
+[project]
 name = "terraform-vault-backend"
-version = "0.3.0"
-description = "A Vault backend for Terraform implemented as an HTTP server."
-authors = ["emanueljg <emanueljohnsongodin@gmail.com>"]
+version = "0.3.1"
+description = "A Vault backend for Terraform implemented as an HTTP server"
+authors = [
+  {name = "Emanuel Johnson Godin", email = "emanueljohnsongodin@gmail.com"},
+]
 readme = "README.md"
-packages = [ {include = "src"}]
+requires-python = ">=3.10"
 
-[tool.poetry.dependencies]
-python = ">=3.9,<4"
-fastapi = ">=0.115.6,<0.116.0"
-hvac = ">=2.3.0,<2.4.0"
-uvicorn = ">=0.34.0,<0.35.0"
-
-
-[tool.poetry.group.dev.dependencies]
-ruff = ">=0.8.5,<0.9.0"
-mypy = ">=1.14.1,<1.15.0"
-pytest = ">=8.3.4,<8.4.0"
-
-[tool.ruff]
-line-length = 100
-fix = true
-exclude = ['stubs']
-
-[tool.ruff.lint]
-select = [ "ALL" ]
-ignore = [
-  "COM812",
-  "ISC001",
-  "D203",
-  "D213",
-  "D104",  
-  "D107",  
-  "ANN401",
-  "FIX002",
-  "TD004",
-  "TD002",
-  "TD003",
-  "FBT001",
-  "ERA001",
-  "PGH003",
-  "RET504",
-  "D301"
+dependencies = [
+    "fastapi>=0.128.0,<0.129.0",
+    "hvac>=2.4.0,<2.5.0",
+    "uvicorn>=0.40.0",  # CVE-2025-43859
+    "h11>=0.16.0",      # fix for CVE-2025-43859
 ]
 
-[tool.ruff.lint.per-file-ignores]
-  "tests/*" = [ "S101" ]
-
-[tool.ruff.lint.flake8-annotations]
-# https://docs.astral.sh/ruff/rules/missing-return-type-special-method/
-mypy-init-return = true 
-  # "D203", "D212", "S101", "EXE002", "D1", "B008" ]
-
+[dependency-groups]
+dev = [
+  "ruff>=0.14.10",
+  "mypy>=1.19.1",
+  "pytest>=9",
+]
 
 [build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+requires = ["hatchling >= 1.26"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/tvb"]
+
